[UPSTREAM?]riscv: Unconditionally check address alignment in SC
Currently, the load is not attempted if the sc address does not match the
last lr address. However, the RISC-V specification says "If the address is
not naturally aligned, an address-misaligned exception or an access-fault
exception will be generated". This change allows qemu to pass TestRIG
atomics passes against sail (100 iterations).