[RVFI-DII] HACK: Only enable wrapping DDC checks when fuzzing
This is quite an ugly hack since we should be doing the checks for wrapping
the end of the address space. However, omitting the checks provides a very
noticeable speedup for code using a full address space DDC. Since the
memory access will fail anyway (due to the wrapping) the only observable
difference is the value of mcause after the trap.