riscv: cheri: address warnings about cpu_restore_state
This is an attempt to tackle warnings about redundant prototypes for
cpu_restore_state.

The order of events is as follows:

target/riscv/vector_helper.c
   target/riscv/cpu.h
      target/cheri-common/cpu_cheri.h
         (hand-coded) prototype cpu_restore_state
   exec/exec-all.h
         second prototype cpu_restore_state

This can be fixed by including exec-all.h from cpu_cheri.h. exec-all.h
will not be included a second time, we have only one prototype.

softmmu/runstate.c
   exec/exec-all.h
      target/riscv/cpu.h
         target/cheri-common/cpu_cheri.h
            exec/exec-all.h -> not included again because of the guard
               cpu_restore_state is used without prototype
      prototype cpu_restore_state

This is more difficult to fix. We can move the cpu_restore_state
prototype in exec-all.h before the line where cpu.h is included.
Not a nice solution, but it works...