cheri: riscv: add the gctag instruction
Cheri Bakewell renamed the cgettag instruction to gctag and modified the
encoding. This instruction reads the tag bit and stores it in a register.

Add support for gctag. Use the same code as for cgettag to read and store
the tag bit. (Generally, we keep the internal representation of the
capability permissions for now, this will be refactored for Bakewell
later.)

The new gctag instruction is supported only for RISC-V. It seems that
Morello is rather different in this area. As for MIPS, our goal is to
keep the current state and make no changes that affect MIPS.