cheri: riscv: remove cgettype
Remove the cgettype instruction for RISC-V. In cheri bakewell, the otype
field was removed from the capability metadata, this makes cgettype
obsolete.

For now, we move the cgettype functionality to mips. We'll have to
disable it completely later when we change the capability encoding (i.e.
when we delete otype).