cheri: risv: remove ccseal
Remove the ccseal instruction for risc-v. It was present in cheri
v9 / p1 and was removed from the specification for cheri bakewell.

Keep the ccseal helper in cheri-common for now. The mips code is still
using it. We might consider moving it to target/mips at some point (it'll
have to be rewritten to not use helpers from cheri-common).