insn32-cheri.decode: rename "cs1 rd" format
We already have a format that contains cs1 and rd registers and we're
about to add a format for "cs1 cd".

Using c2 as a format name would be ambiguous. Rename the "cs1 rd" format
to cr (the "cs1 cd" format will then be called cc).

While at it, reformat and reorder some definitions.