cheri: riscv: add the sentry instruction
Cheri Bakewell renamed the CSealEntry instruction to sentry and modified the
encoding. This instruction makes a sealed copy of a capability.

This patch adds support for sentry. Like other bakewell instructions,
sentry is supported only for RISC-V.

The function to seal a capability requires an otype for cheri v9/p1, but
not for bakewell. Use a dummy define to make this clear.