cheri: riscv: add the cbld instruction
Cheri Bakewell renamed the cbuildcap instruction to cbld and modified the
encoding. This instruction takes an authorizing capability and an input
capability. If permitted, a new tagged capability is created from the
input capability. It seems that the related permission checks were
modfied for cheri bakewell.

Reimplement cbld based on the v9 code. Make sure that the checks are in
line with the bakewell specification. Execute cbld only in capability
pointer mode.

For the required malformed checks, we assume that bounds length 0 is
equivalent to a malformed capability.