cheri: remove cheri_compressed_cap.h from qemu build
The cheri version of qemu used the cheri-compressed-cap library to handle
compressed capabilities.

cheri-compressed-cap also supports a 256-bit uncompressed representation
of a capability in memory. It seems that this 256-bit format was defined
in early cheri specifications and has been removed some time before V9/P1.

This 256-bit uncompressed format is defined in cheri_compressed_cap.h.
Although qemu does not use this format, it includes cheri_compressed_cap.h.
The only code from this file that qemu needs is further include statements
for the 64-bit, 128-bit and 128-bit morello representations of a capability.

This patch modifies qemu to include the three variants directly and
remove cheri_compressed_cap.h from the qemu build. We can then ignore the
unused 256-bit format when we update the capability format for cheri bakewell.

(Generally, we try to keep the changes to the cheri-compressed-cap in
qemu to a minimum. We might have to change the rest of the library for
cheri bakewell as a separate project later.)