cheri: riscv: fix processing of cadd with increment 0 (cmv)
cadd cd, cs1, x0 can be written as cmv cd, cs1. cs1 is copied to cd. In
this special case, copying a sealed register is supported. cd's tag is
not cleared if cs1 is sealed.

Handle this case directly in trans_cadd. Create tcg instructions to copy
the capability register.