cheri-compressed-cap: fix bounds encoding for 64-bit bakewell
Fix the encoding of base, length and top for cheri bakewell with
MXLEN=32. In addition to EF, E, B[9:0] and T[7:0], we also have to
store bit 8 of the length in the L8 field.

Add the L8 field and explain in detail why L8 is needed for recovering
T[9:8].

This issue was analyzed and fixed by Paul, all credit goes to him.