cheri: riscv: illegal instruction for gcperm in int ptr mode
The gcperm instruction is allowed only in purecap mode or in hybrid mode
when the system is in capability pointer mode.

Raise an illegal instruction exception if gcperm is called in hybrid,
integer pointer mode.

The gcperm instruction is generic, the exception is riscv-specific. It
seems that we have to bite the bullet and use ifdef here...