cheri: riscv: use correct mode for exception handlers
Use the correct (capability or integer pointer) mode when we jump to an
exception handler.

When an exception is raised, the code checks the M bit in PCC to
determine the current mode. To jump into the exception handler, it
installs either the address in xTVEC or the capability in xTVECC as the
new PCC.

This does not work if the exception was caused by an invalid PCC. In this
case, we must not use this invalid PCC for setting up the exception
handler.

Rewrite the mode check to use the xTVECC CSR. If xTVECC contains a valid
capability that grants execute permission, we can safely assume that the
exception handler should run in capability pointer mode.

Fixes: 2cc3ba325a4 ("riscv: fix pc / pcc update for trap handling")