cheri: define to check for the risc-v formats
Add a define for checking if we're using one of the two risc-v capability
formats (also known as bakewell internally). We should not link this to
TARGET_RISCV. We might have to support other formats on risc-v and the
risc-v formats might be used on other architectures in future.

Update the format check in cheri_utils.h to use the new define.