riscv: cheri: remove cap load-reserved authorized by other cap
The current cheri risc-v specification does no longer define a separate
instruction for capability load-reserved authorized by another
capability. Such an instruction was present in the v9 specification.

Remove the encoding for this instruction and its translation function.
Remove the helper function as well, it is used only for risc-v.