hw/riscv/hobgoblin: Rework Genesys2/proFPGA code to use qemu object model
This was motivated by adding smp support to the proFPGA platform.
The order that qemu initialises itself means that the -smp command
line option is parsed before the machine properties are set, so
checks against the maximum number of cores would happen before knowing
whether this is a Genesys 2 or proFPGA.

So rework the code following the example of the Arm Musca-B1 (although
many other ARM boards use the same technique):
 - make the hobgoblin type an abstract type
 - derive two new types: hobgoblin-genesys2 and hobgoblin-profpga
 - the board type become a member of the hobgloblin class rather
   than the instance of the derived class.

Also switch to using OBJECT_DECLARE_TYPE(), which defines helper macros
to cast between object types, but also some typedefs, which does mean
that the names of the types no longer have _t postfix.

Signed-off-by: Stuart Menefy <stuart.menefy@codasip.com>