hw: misc: Add Codasip TRNG (True Random Number Generator)
This is a simple emulation of the Codasip TRNG (True Random Number
Generator), based on the documentation (version 0.01). There is
no distinction between the raw and processed FIFOs, and the FIFO
is only one element deep, but these differences from the real hardware
shouldn't be visible to most users. The reset value of the CONFIG
register is unknown at this point, but as it only affects parts of the
low level behavior of hardware which aren't visible to the user this
shouldn't matter.

Signed-off-by: Stuart Menefy <stuart.menefy@codasip.com>