configs: riscv64cheri: remove the sifive e board
Drop support for the "sifive e" board for riscv64-cheri-softmmu.

This board has 0x4000 bytes of ram. Now that the ram area is configured
as cheri tagmem automatically, this triggers an assert. The tagmem code
allocates blocks of 4096 tag bits. And it requires that all bits of such
a block be linked to actual ram addresses. For RV64, we'd need 4096 * 16
bytes == 64k RAM to use all tags of a tagblock....

At the moment, riscv64cheri-softmmu/default.mak includes
riscv64-softmmu/default.mak. It seems that we can't include a config and
then overwrite some of its settings.

We have to copy the settings and remove sifive e.