cheri-compressed-cap: update the subrepo
cheri-compressed-cap is a standalone library that is part of the qemu
tree. It is maintained as a git subrepo.

With git subrepo, users and developers of the qemu repository do not need
access to the cheri-compressed-cap subrepo. The qemu repository contains
a copy that is synced regularly.

This commit updates the subrepo configuration to point to Codasip's
internal cheri-compressed-cap repository.