Update the RV32 SC instruction to use the same encoding as RV64
THe 0.90 spectification unifies the encodings for the SC instruction so that both RV32 and RV64 use the same (RV64) encoding.
Remove the RV32 varient.
Move the Unified version out of the ins32-cheri-64.decode file into the shared inst32-cheri.decode.
Fix the instruction to use the correct predicate (pred_cre)
Update the disassembler.