Unify encoding of lr_c instruction.
The Updateed riscv spec unifies the encodings of the RV32 and RV64 encodings of lr.c to use the RV64 encoding.
Move the RV64 encoding out of the RV64 specific file into the common file.
Remove the RV32 decoing.
Remove the trans_stub.
Add pred_cre for the instruction
This instruction does not appear to have a disassebler implemented. Add this later