Fix an issue where guest code could cause a qemu crash.
When asked to invalidate tags outside the allocated ram range we could end up generating an illegal tagblock address. Replace the debug assert that protects this in debug mode with a NULL ptr return, and a error message.
Most callers already checked for the nullptr and attempt to handle. For the remaining case add a debug_assert.